+++ packages/cli/CHANGELOG.md
@@ -1,5 +1,76 @@
 # aeria-build
 
+## 0.0.41
+
+### Patch Changes
+
+- Updated dependencies [afbd76e]
+  - @aeriajs/common@0.0.63
+  - @aeriajs/builtins@0.0.102
+  - @aeriajs/core@0.0.102
+  - @aeriajs/entrypoint@0.0.65
+
+## 0.0.40
+
+### Patch Changes
+
+- Updated dependencies [535b932]
+  - @aeriajs/builtins@0.0.101
+  - @aeriajs/core@0.0.101
+
+## 0.0.39
+
+### Patch Changes
+
+- 5e6751d: Change `describe()` signature
+- Updated dependencies [5e6751d]
+  - @aeriajs/builtins@0.0.100
+  - @aeriajs/core@0.0.100
+
+## 0.0.38
+
+### Patch Changes
+
+- Updated dependencies [4c7243e]
+  - @aeriajs/entrypoint@0.0.64
+  - @aeriajs/builtins@0.0.99
+  - @aeriajs/core@0.0.99
+
+## 0.0.37
+
+### Patch Changes
+
+- Updated dependencies [ebe7477]
+  - @aeriajs/types@0.0.58
+  - @aeriajs/builtins@0.0.98
+  - @aeriajs/common@0.0.62
+  - @aeriajs/core@0.0.98
+  - @aeriajs/entrypoint@0.0.63
+
+## 0.0.36
+
+### Patch Changes
+
+- Updated dependencies [6c981e6]
+- Updated dependencies [ec235f7]
+  - @aeriajs/common@0.0.61
+  - @aeriajs/types@0.0.57
+  - @aeriajs/builtins@0.0.97
+  - @aeriajs/core@0.0.97
+  - @aeriajs/entrypoint@0.0.62
+
+## 0.0.35
+
+### Patch Changes
+
+- fafac8a: Introduce `Result`
+- Updated dependencies [fafac8a]
+  - @aeriajs/builtins@0.0.96
+  - @aeriajs/common@0.0.60
+  - @aeriajs/types@0.0.56
+  - @aeriajs/core@0.0.96
+  - @aeriajs/entrypoint@0.0.61
+
 ## 0.0.34
 
 ### Patch Changes