+++ packages/types/CHANGELOG.md
@@ -1,5 +1,23 @@
 # @aeriajs/types
 
+## 0.0.58
+
+### Patch Changes
+
+- ebe7477: Fix `FixedObjectProperty` type
+
+## 0.0.57
+
+### Patch Changes
+
+- ec235f7: Fix pagination and function signatures
+
+## 0.0.56
+
+### Patch Changes
+
+- fafac8a: Introduce `Result`
+
 ## 0.0.55
 
 ### Patch Changes